v0.1/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
jinja2>=3.1
python-multipart>=0.0.9
python-dotenv>=1.0
httpx>=0.27
msal>=1.28
